Fix protocol error with https in front of site.url
According to the `_config.yml` file, the `url` variable
should include the protocol (e.g. `https://`). However,
at other places thoughout the code, you used
`https://{{ site.url }}`, which lead to some wierd issues
when hosting with a custom domain on github (mainly issues
where some resources are requested using http instead of https).

I hope this is the way `site.url` was intended. We would have
to agree on where to put the protocol. For now, it has to be
added to the `url` variable. Thus, we have to drop any `https://`
in front of `{{ site.url }}`.